REFACTOR mini fix (✓E2E ✓UNIT)
