+++ src/features/app/layout/_components/CookieBar.UNUSED.tsx
